Upload current status of my MegaEdit practice script
Started working on creating a Name Tag template with MegaEdit for the ASU Print and Imaging Lab. This is a practice enviornment. So far it has three text fields with Name, Job Title and Description. When you begin typing in the text box for each of these three fields it will appropriately update the canvas to the letters the user inputs. I also have a dropdown created to change the size of the name text depending on if the user selects Small, Medium, or Large. Lastly, I created a dropdown to insert a specific logo that the user wants in the top left hand corner of the screen.